Revert "v3.29.1 (#3)"
This reverts commit 2361681384947103094930fed29b89673511fdeb.

+++ drivers/123/upload.go
@@ -107,7 +107,7 @@ func (d *Pan123) newUpload(ctx context.Context, upReq *UploadResp, file model.Fi
 			if err != nil {
 				return err
 			}
-			up(float64(j) * 100 / float64(chunkCount))
+			up(j * 100 / chunkCount)
 		}
 	}
 	// complete s3 upload
